requirements.txt
livekit-agents[silero,deepgram,groq,turn-detector]>=1.3.0,<2.0.0
livekit-api>=1.1.0
fastapi>=0.110.0
uvicorn>=0.29.0
plivo>=4.47.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
loguru>=0.7.0
